Removed Travis & Reformatted Tests
1. Removed Travis. Do in next PR.
2. Renamed tests for WPCS.
