--- .travis.yml
@@ -1,44 +1,0 @@
-# Travis CI (MIT license) configuration file for Beans.
-# @link https://travis-ci.org/
-
-# For use with the Beans Theme Framework
-# @link https://github.com/Getbeans/Beans
-
-# This version does not include sniffers. Why? The codebase would fail.
-# Once we get the codebase compliant, then we'll update this file.
-
-sudo: false
-dist: trusty
-language: php
-
-env:
-  global:
-    - WP_VERSION=master WP_MULTISITE=0
-
-matrix:
-    fast_finish: true
-
-    include:
-# Disabling checks on 5.2 for now as tests will run at 5.6.
-#        - php: 5.2
-#          dist: precise
-        - php: 5.6
-        - php: 7.0
-        - php: 7.2
-
-    allow_failures:
-        - php: 7.2
-
-before_install:
-    - composer self-update
-
-install:
-  - composer install --no-interaction
-
-script:
-  - composer test
-
-notifications:
-  email:
-    on_success: never
-    on_failure: change
